rocket-calc, a small tool to help design rockets
Uses a DFS search to iterate through rocket possiblities and find a
good solution to the rocket problem; that is, what is a "Good" design
for a rocket. Redesigns upper stages if lower stages have enough
energy to handle the problem, thus making a smaller and lighter
spacecraft overall. You might need to do two iterations for more
complex missions, e.g. an Apollo-style lander-orbiter mission.